OmajauTheChatBot/Agent/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
openai==0.28.1
Werkzeug==2.3.7
requests==2.31.0
SpeechRecognition==3.10.0
gTTS==2.3.2
pyttsx3==2.90
PyAudio==0.2.11
